Make it possible to disable proximity sensor during calls
There are phones out there with broken proximity sensors that always
report near, which causes every dialer app to turn off the screen
unconditionally during calls. As an owner of such a phone, this has been
driving me crazy.

This commit adds a new setting to disable the proximity sensor. By
default, this is turned off, preserving the current behaviour.